Fixed module uninstallation. Fixes #18

+++ src/AI4E.Modularity.Host/ModuleInstallationManager.cs
@@ -1,3 +1,23 @@
+/* License
+ * --------------------------------------------------------------------------------------------------------------------
+ * This file is part of the AI4E distribution.
+ *   (https://github.com/AI4E/AI4E)
+ * Copyright (c) 2018 - 2019 Andreas Truetschel and contributors.
+ * 
+ * AI4E is free software: you can redistribute it and/or modify  
+ * it under the terms of the GNU Lesser General Public License as   
+ * published by the Free Software Foundation, version 3.
+ *
+ * AI4E is distributed in the hope that it will be useful, but 
+ * WITHOUT ANY WARRANTY; without even the implied warranty of 
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU 
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ * --------------------------------------------------------------------------------------------------------------------
+ */
+
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -143,20 +163,30 @@ namespace AI4E.Modularity.Host
             await supervisor.DisposeAsync();
             _supervisors.Remove(supervisor);
 
-            while (cancellation.ThrowOrContinue())
+            while (!DeleteModuleDirectory(moduleInstallationDirectory))
             {
-                try
+                cancellation.ThrowIfCancellationRequested();
+                await Task.Delay(1000);
+            }
+        }
+
+        private bool DeleteModuleDirectory(DirectoryInfo moduleInstallationDirectory)
+        {
+            try
+            {
+                if (moduleInstallationDirectory.Exists)
                 {
                     moduleInstallationDirectory.Delete(recursive: true);
                 }
-                catch (DirectoryNotFoundException) { }
-                catch
-                {
-                    // TODO: Log warning
-
-                    await Task.Delay(1000);
-                }
             }
+            catch (DirectoryNotFoundException) { }
+            catch
+            {
+                // TODO: Log error
+                return false;
+            }
+
+            return true;
         }
 
         private async Task<IDictionary<ModuleReleaseIdentifier, IModuleSupervisor>> GetSupervisorsAsync(CancellationToken cancellation)
